requirements.txt
faker
openai
pandas
pydantic
python-dotenv
requests
tqdm